moved helper functions for better code readability
