+++ test_approach/new_pipeline_trajectory_generation.py
@@ -18,6 +18,16 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from utils.new_generate_trajectory import chat_ai_playwright_code
 from config import RESULTS_DIR, ACCOUNTS, BROWSER_SESSIONS_DIR
 from utils.google_auth import ensure_google_login
+from utils.trajectory_file_utils import (
+    create_episode_directory, create_trajectory_file, create_error_log_file,
+    update_playwright_error_log, update_trajectory, create_metadata,
+    write_user_message, generate_trajectory_html
+)
+from utils.element_utils import (
+    get_comprehensive_element_data, create_simplified_element_summary,
+    try_alternative_selectors, annotate_screenshot_with_bounding_boxes,
+    get_all_open_tabs, check_for_new_tabs, switch_to_new_tab
+)
 
 # Load environment variables from .env file
 load_dotenv()
@@ -25,644 +35,6 @@ load_dotenv()
 # Knowledge base client for trajectory context
 from utils.knowledge_base_client import get_trajectory_context
 
-
-
-def get_comprehensive_element_data(page, url: str = None) -> Dict[str, Any]:
-    """
-    Get interactive elements directly using Playwright - simplified approach.
-    
-    Args:
-        page: Playwright page object
-        url: Current page URL for context-specific filtering
-        
-    Returns:
-        Dict containing interactive elements and targeting data
-    """
-    print("🔍 Collecting interactive elements...")
-    
-    # Get all interactive elements directly using Playwright
-    ax_elements = get_all_interactive_elements(page)
-    print(f"    Found {len(ax_elements)} interactive elements")
-    
-    # Create targeting data
-    targeting_data = create_comprehensive_targeting_data(ax_elements, url)
-    
-    return {
-        "interactive_elements": ax_elements,
-        "targeting_data": targeting_data,
-        "element_count": len(ax_elements),
-        "collection_timestamp": time.time()
-    }
-
-def create_simplified_element_summary(targeting_data: list) -> list:
-    """
-    Create simplified element data with just annotation_id, role, and name.
-    This is what gets sent to GPT and saved in axtree files.
-    
-    Args:
-        targeting_data: Full targeting data from comprehensive element collection
-        
-    Returns:
-        List of simplified element objects
-    """
-    elements_data = []
-    for elem in targeting_data:
-        annotation_id = elem.get('annotation_id', '?')
-        role = elem.get('element_info', {}).get('role', 'unknown')
-        name = elem.get('element_info', {}).get('name', 'unnamed')
-        elements_data.append({
-            "annotation_id": str(annotation_id),
-            "role": role,
-            "name": name
-        })
-    return elements_data
-
-
-
-
-
-
-
-def get_all_interactive_elements(page) -> list:
-    """Get all interactive elements directly using Playwright - simple and fast"""
-    elements = []
-    
-    # List of interactive roles to look for
-    interactive_roles = [
-        'button', 'link', 'textbox', 'checkbox', 'radio', 'combobox',
-        'listbox', 'menuitem', 'tab', 'slider', 'spinbutton', 'searchbox',
-        'switch', 'menubar', 'toolbar', 'tree', 'grid', 'table',
-        'option', 'menuitemcheckbox', 'menuitemradio', 'listitem',
-        'group', 'region', 'dialog', 'alertdialog', 'tooltip'
-    ]
-
-    # Get elements by each role
-    for role in interactive_roles:
-        try:
-            role_elements = page.get_by_role(role).all()
-            for element in role_elements:
-                try:
-                    bbox = element.bounding_box()
-                    # Filter out elements with negative coordinates or very small dimensions
-                    # if bbox and bbox['width'] > 2 and bbox['height'] > 2 and bbox['x'] >= 0 and bbox['y'] >= 0:
-                    if bbox and bbox['width'] > 8 and bbox['height'] > 8 and bbox['x'] >= 0 and bbox['y'] >= 0:
-                        # Check if element is truly visible and not transparent/covered
-                        try:
-                            if not element.is_visible():
-                                continue
-                            
-                            # Additional check: ensure element has actual visual presence and is in viewport
-                            is_visually_visible = element.evaluate('''
-                                (el) => {
-                                    const style = window.getComputedStyle(el);
-                                    const rect = el.getBoundingClientRect();
-                                    const viewport = {
-                                        width: window.innerWidth,
-                                        height: window.innerHeight
-                                    };
-                                    
-                                    // Check if element has no dimensions
-                                    if (rect.width === 0 || rect.height === 0) return false;
-                                    
-                                    // Check if element is transparent
-                                    if (parseFloat(style.opacity) === 0) return false;
-                                    
-                                    // Check if element is hidden by CSS
-                                    if (style.display === 'none' || style.visibility === 'hidden') return false;
-                                    
-                                    // Check if element is hidden by CSS animations or transitions
-                                    if (style.animationName !== 'none' && style.animationPlayState === 'paused') return false;
-                                    if (style.transitionDuration !== '0s' && style.transitionProperty !== 'none') {
-                                        // If element is in transition, check if it's actually visible
-                                        if (parseFloat(style.opacity) < 0.1) return false;
-                                    }
-                                    
-                                    // Check if element is positioned off-screen or partially hidden
-                                    if (rect.right < 0 || rect.bottom < 0 || rect.left > viewport.width || rect.top > viewport.height) return false;
-                                    
-                                    // Check if element is completely outside the viewport (more strict)
-                                    if (rect.left >= viewport.width || rect.top >= viewport.height) return false;
-                                    
-                                    // Check if element is too small to be meaningful (less than 8x8 pixels)
-                                    if (rect.width < 8 || rect.height < 8) return false;
-                                    
-                                    // Check if element is covered by another element (z-index issues)
-                                    const elementAtPoint = document.elementFromPoint(
-                                        rect.left + rect.width / 2,
-                                        rect.top + rect.height / 2
-                                    );
-                                    if (!elementAtPoint || !el.contains(elementAtPoint)) return false;
-                                    
-                                    // Additional check: ensure element is not clipped by overflow
-                                    const parent = el.parentElement;
-                                    if (parent) {
-                                        const parentStyle = window.getComputedStyle(parent);
-                                        if (parentStyle.overflow === 'hidden' || parentStyle.overflow === 'scroll') {
-                                            const parentRect = parent.getBoundingClientRect();
-                                            if (rect.left < parentRect.left || rect.top < parentRect.top || 
-                                                rect.right > parentRect.right || rect.bottom > parentRect.bottom) {
-                                                return false;
-                                            }
-                                        }
-                                    }
-                                    
-                                    // Check if element has no meaningful content (text, images, or interactive properties)
-                                    const hasText = el.textContent && el.textContent.trim().length > 0;
-                                    const hasImage = el.querySelector('img') || el.tagName === 'IMG';
-                                    const hasInteractiveAttr = el.getAttribute('onclick') || el.getAttribute('href') || el.getAttribute('role') || el.tagName === 'BUTTON' || el.tagName === 'INPUT' || el.tagName === 'A';
-                                    const hasBackground = style.backgroundImage !== 'none' || style.backgroundColor !== 'rgba(0, 0, 0, 0)';
-                                    
-                                    // Element must have at least one of these to be considered visible
-                                    if (!hasText && !hasImage && !hasInteractiveAttr && !hasBackground) return false;
-                                    
-                                    // Final check: ensure element is actually in the current scroll viewport
-                                    const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
-                                    const scrollLeft = window.pageXOffset || document.documentElement.scrollLeft;
-                                    const viewportTop = scrollTop;
-                                    const viewportBottom = scrollTop + viewport.height;
-                                    const viewportLeft = scrollLeft;
-                                    const viewportRight = scrollLeft + viewport.width;
-                                    
-                                    if (rect.top > viewportBottom || rect.bottom < viewportTop || 
-                                        rect.left > viewportRight || rect.right < viewportLeft) {
-                                        return false;
-                                    }
-                                    
-                                    // Check if element is actually clickable/interactive in current state
-                                    const isClickable = el.offsetParent !== null && 
-                                                       !el.disabled && 
-                                                       el.style.pointerEvents !== 'none' &&
-                                                       el.style.userSelect !== 'none';
-                                    
-                                    if (!isClickable) return false;
-                                    
-                                    return true;
-                                }
-                            ''')
-                            
-                            if not is_visually_visible:
-                                continue
-                        except:
-                            continue
-                        # Get element properties
-                        name = element.get_attribute('aria-label') or \
-                               element.text_content() or \
-                               element.get_attribute('title') or \
-                               element.get_attribute('placeholder') or \
-                               element.get_attribute('value') or ''
-                        
-                        tag_name = element.evaluate('el => el.tagName.toLowerCase()')
-                        element_id = element.get_attribute('id') or ''
-                        class_name = element.get_attribute('class') or ''
-                        href = element.get_attribute('href') or ''
-                        element_type = element.get_attribute('type') or ''
-                        disabled = element.evaluate('el => el.disabled') or False
-                        checked = element.evaluate('el => el.checked')
-                        selected = element.evaluate('el => el.selected')
-                        
-                        # Create clean Playwright selector without action (action will be determined by GPT's action_type)
-                        playwright_selector = None
-                        if element_id:
-                            playwright_selector = f"page.locator('#{element_id}')"
-                        elif name.strip():
-                            playwright_selector = f'page.get_by_text("{name.strip()}")'
-                        else:
-                            playwright_selector = f'page.get_by_role("{role}")'
-                        
-
-                        
-                        element_data = {
-                            'name': name.strip() if name else '',
-                            'role': role,
-                            'value': element.get_attribute('value') or '',
-                            'x': int(bbox['x']),
-                            'y': int(bbox['y']),
-                            'width': int(bbox['width']),
-                            'height': int(bbox['height']),
-                            'tagName': tag_name,
-                            'type': element_type,
-                            'id': element_id,
-                            'className': class_name,
-                            'href': href,
-                            'disabled': disabled,
-                            'checked': checked,
-                            'selected': selected,
-                            'source': 'playwright_direct',
-                            'hasBoundingBox': True,
-                            'playwright_selector': playwright_selector
-                        }
-                        elements.append(element_data)
-                except:
-                    continue
-        except:
-            continue
-    
-    return elements
-
-
-
-def create_comprehensive_targeting_data(elements: list, url: str = None) -> list:
-    """Create comprehensive targeting data for elements with multiple strategies"""
-    targeting_data = []
-    
-    for i, element in enumerate(elements):
-        # Create multiple targeting strategies for each element
-        element_data = {
-            "annotation_id": i,
-            "element_info": {
-                "name": clean_text_for_selector(element.get('name', '')),
-                "name_raw": element.get('name', ''),  # Keep raw version too
-                "role": element.get('role', ''),
-                "value": element.get('value', ''),
-                "tag_name": element.get('tagName', ''),
-                "type": element.get('type', ''),
-                "id": element.get('id', ''),
-                "class_name": element.get('className', ''),
-                "href": element.get('href', ''),
-                "disabled": element.get('disabled', False),
-                "checked": element.get('checked'),
-                "selected": element.get('selected')
-            },
-            "bounding_box": {
-                "x": element.get('x', 0),
-                "y": element.get('y', 0),
-                "width": element.get('width', 0),
-                "height": element.get('height', 0),
-                "center_x": element.get('x', 0) + element.get('width', 0) // 2,
-                "center_y": element.get('y', 0) + element.get('height', 0) // 2
-            },
-            "playwright_selectors": generate_playwright_selectors(element),
-            "interaction_suggestions": suggest_interactions(element)
-        }
-        
-        targeting_data.append(element_data)
-    
-    return targeting_data
-
-def clean_text_for_selector(text: str) -> str:
-    """Clean text for use in selectors - conservative cleaning"""
-    if not text:
-        return ""
-    
-    # Only remove leading/trailing whitespace and normalize newlines
-    # Keep internal spacing as-is to preserve legitimate spaces
-    cleaned = text.strip()
-    
-    # Replace newlines and tabs with single spaces, but preserve regular spaces
-    cleaned = cleaned.replace('\n', ' ').replace('\t', ' ').replace('\r', ' ')
-    
-    # Only collapse multiple spaces if there are 3+ consecutive spaces
-    # This preserves intentional double spaces but removes excessive whitespace
-    cleaned = re.sub(r' {3,}', ' ', cleaned)
-    
-    # Escape single quotes for selector strings
-    cleaned = cleaned.replace("'", "\\'")
-    
-    # Truncate if too long
-    if len(cleaned) > 50:
-        cleaned = cleaned[:47] + "..."
-    
-    return cleaned
-
-def try_alternative_selectors(page, original_code: str, comprehensive_data: dict, gpt_resp: dict) -> tuple[bool, list, str]:
-    """
-    Try alternative Playwright selectors when the primary one fails.
-    Simple approach: loop through selectors array (skip first) and try exec() one by one.
-    Returns: (success, failed_alternatives, successful_selector_code)
-    """
-    failed_alternatives = []
-    
-    try:
-        # Get the selected annotation ID from GPT response
-        selected_id = gpt_resp.get('selected_annotation_id')
-        if not selected_id:
-            print("⚠️ No selected_annotation_id found, can't try alternative selectors")
-            return False, failed_alternatives, ""
-        
-        # Find the element data for this annotation ID
-        target_element = None
-        for element in comprehensive_data.get('targeting_data', []):
-            if str(element.get('annotation_id', '')) == str(selected_id):
-                target_element = element
-                break
-        
-        if not target_element:
-            print(f"⚠️ Element with annotation ID {selected_id} not found in targeting data")
-            return False, failed_alternatives, ""
-        
-        # Get alternative selectors
-        alternative_selectors = target_element.get('playwright_selectors', [])
-        if len(alternative_selectors) == 0:
-            print("⚠️ No alternative selectors available")
-            return False, failed_alternatives, ""
-        
-        print(f"🎯 Trying alternative selectors for element {selected_id}")
-        
-        # Try each alternative selector starting from index 0
-        for i, selector_data in enumerate(alternative_selectors):
-            selector_code = selector_data.get('selector', '')
-            selector_type = selector_data.get('type', 'unknown')
-            
-            if not selector_code:
-                continue
-            
-            # Construct the action based on the action type from GPT's response
-            # Check if the selector already includes an action (like coordinates or mouse actions)
-            action_type = gpt_resp.get('action_type', 'click')  # Default to click if not specified
-            
-            # Check if selector already includes an action
-            if 'page.mouse.click' in selector_code or 'page.keyboard.' in selector_code:
-                # Selector already includes the action, use as-is
-                if action_type == 'fill':
-                    text_to_fill = gpt_resp.get('text_to_fill', 'TEXT_TO_FILL')
-                    action_code = f"{selector_code}; page.keyboard.type('{text_to_fill}')"
-                else:
-                    action_code = selector_code
-            else:
-                # Selector is clean, add the appropriate action
-                if action_type == 'click':
-                    action_code = f"{selector_code}.click()"
-                elif action_type == 'fill':
-                    # For fill actions, use the text_to_fill from GPT's response if available
-                    text_to_fill = gpt_resp.get('text_to_fill', 'TEXT_TO_FILL')
-                    action_code = f"{selector_code}.fill('{text_to_fill}')"
-                elif action_type == 'select':
-                    action_code = f"{selector_code}.select_option('OPTION_TO_SELECT')"
-                elif action_type == 'navigate':
-                    action_code = f"{selector_code}.click()"  # Navigate usually involves clicking
-                elif action_type == 'wait':
-                    action_code = f"{selector_code}.wait_for()"  # Wait for element to be visible
-                else:
-                    # Default to click for unknown action types
-                    action_code = f"{selector_code}.click()"
-            
-            print(f"  🔄 Trying alternative {i}: {selector_type} - {action_code}")
-            
-            try:
-                # Execute the alternative selector with the constructed action
-                exec(action_code)
-                print(f"  ✅ Alternative selector succeeded: {selector_type}")
-                return True, failed_alternatives, action_code
-                
-            except Exception as alt_e:
-                print(f"  ❌ Alternative {i} failed: {alt_e}")
-                failed_alternatives.append(action_code)
-                continue
-        
-        print("❌ All alternative selectors failed")
-        return False, failed_alternatives, ""
-        
-    except Exception as e:
-        print(f"⚠️ Error trying alternative selectors: {e}")
-        return False, failed_alternatives, ""
-
-def generate_colors(count):
-    """Generate distinct colors for bounding boxes"""
-    colors = []
-    for i in range(count):
-        # Generate bright colors that are easily distinguishable
-        hue = (i * 137.508) % 360  # Golden angle for good distribution
-        saturation = 70 + (i % 3) * 10  # 70, 80, 90
-        lightness = 50 + (i % 2) * 20   # 50, 70
-        
-        # Convert HSL to RGB
-        c = (1 - abs(2 * lightness/100 - 1)) * saturation/100
-        x = c * (1 - abs((hue / 60) % 2 - 1))
-        m = lightness/100 - c/2
-        
-        if 0 <= hue < 60:
-            r, g, b = c, x, 0
-        elif 60 <= hue < 120:
-            r, g, b = x, c, 0
-        elif 120 <= hue < 180:
-            r, g, b = 0, c, x
-        elif 180 <= hue < 240:
-            r, g, b = 0, x, c
-        elif 240 <= hue < 300:
-            r, g, b = x, 0, c
-        else:
-            r, g, b = c, 0, x
-        
-        r = int((r + m) * 255)
-        g = int((g + m) * 255)
-        b = int((b + m) * 255)
-        
-        colors.append((r, g, b))
-    
-    return colors
-
-def annotate_screenshot_with_bounding_boxes(screenshot_path: str, targeting_data: list, annotated_path: str) -> str:
-    """
-    Annotate screenshot with bounding boxes and annotation IDs for interactive elements.
-    
-    Args:
-        screenshot_path: Path to the original screenshot
-        targeting_data: List of element data with bounding boxes and annotation IDs
-        annotated_path: Path to save the annotated image
-        
-    Returns:
-        Path to the annotated image
-    """
-    try:
-        from PIL import Image, ImageDraw, ImageFont
-        
-        # Open the screenshot
-        img = Image.open(screenshot_path)
-        draw = ImageDraw.Draw(img)
-        
-        # Try to load a font, fallback to default if not available
-        try:
-            font = ImageFont.truetype("arial.ttf", 16)
-        except:
-            try:
-                font = ImageFont.truetype("/System/Library/Fonts/Arial.ttf", 16)  # macOS
-            except:
-                try:
-                    font = ImageFont.truetype("/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf", 16)  # Linux
-                except:
-                    font = ImageFont.load_default()
-        
-        # Generate colors for each element
-        colors = generate_colors(len(targeting_data))
-        
-        # Draw bounding boxes and labels
-        for i, element in enumerate(targeting_data):
-            bbox = element.get('bounding_box', {})
-            annotation_id = element.get('annotation_id', '?')
-            
-            x = bbox.get('x', 0)
-            y = bbox.get('y', 0)
-            width = bbox.get('width', 0)
-            height = bbox.get('height', 0)
-            
-            # Skip elements with invalid dimensions
-            if width <= 0 or height <= 0:
-                continue
-            
-            color = colors[i]
-            
-            # Draw bounding box
-            draw.rectangle([x, y, x + width, y + height], outline=color, width=2)
-            
-            # Create label text (just the annotation ID)
-            label = f"{annotation_id}"
-            
-            # Draw label background
-            text_bbox = draw.textbbox((0, 0), label, font=font)
-            text_width = text_bbox[2] - text_bbox[0]
-            text_height = text_bbox[3] - text_bbox[1]
-            
-            label_x = max(0, min(x, img.width - text_width - 4))
-            label_y = max(0, y - text_height - 4)
-            
-            draw.rectangle(
-                [label_x, label_y, label_x + text_width + 4, label_y + text_height + 4],
-                fill=color,
-                outline=color
-            )
-            
-            # Draw label text (white for better visibility)
-            draw.text(
-                (label_x + 2, label_y + 2),
-                label,
-                fill='white',
-                font=font
-            )
-            
-            # # Draw click coordinate marker (center point of the element)
-            # click_x = x + width // 2
-            # click_y = y + height // 2
-            
-            # # Draw a small circle at the click coordinates
-            # circle_radius = 3
-            # draw.ellipse(
-            #     [click_x - circle_radius, click_y - circle_radius, 
-            #      click_x + circle_radius, click_y + circle_radius],
-            #     fill='yellow',
-            #     outline='black',
-            #     width=1
-            # )
-            
-
-        
-        # Save the annotated image
-        img.save(annotated_path)
-        print(f"✅ Annotated screenshot saved to: {annotated_path}")
-        
-        return annotated_path
-        
-    except ImportError:
-        print("⚠️ PIL not available, skipping image annotation")
-        return screenshot_path
-    except Exception as e:
-        print(f"⚠️ Error annotating screenshot: {e}")
-        return screenshot_path
-
-def generate_playwright_selectors(element: dict) -> list:
-    """Generate multiple Playwright selector strategies for an element"""
-    selectors = []
-    
-    # Strategy 1: By ID (highest priority - most reliable)
-    if element.get('id'):
-        selectors.append({
-            "type": "id",
-            "selector": f"page.locator('#{element['id']}')",
-            "priority": "high"
-        })
-    
-    # Strategy 2: By coordinates (mouse click) - second priority
-    selectors.append({
-        "type": "coordinates",
-        "selector": f"page.mouse.click({element.get('x', 0) + element.get('width', 0)//2}, {element.get('y', 0) + element.get('height', 0)//2})",
-        "priority": "high"
-    })
-    
-    # Strategy 3: By role and name
-    if element.get('role') and element.get('name'):
-        selectors.append({
-            "type": "role_name",
-            "selector": f"page.get_by_role('{element['role']}', name='{element['name']}')",
-            "priority": "high"
-        })
-    
-    # Strategy 4: By label
-    if element.get('name'):
-        selectors.append({
-            "type": "label",
-            "selector": f"page.get_by_label('{element['name']}')",
-            "priority": "high"
-        })
-    
-    # Strategy 5: By text content
-    if element.get('name'):
-        selectors.append({
-            "type": "text",
-            "selector": f"page.get_by_text('{element['name']}')",
-            "priority": "medium"
-        })
-    
-    # Strategy 6: By CSS class combination
-    css_parts = []
-    if element.get('tagName'):
-        css_parts.append(element['tagName'])
-    if element.get('id'):
-        css_parts.append(f"#{element['id']}")
-    if element.get('className'):
-        # Split class names and add them
-        classes = element['className'].split()
-        for cls in classes[:3]:  # Limit to first 3 classes
-            if cls.strip():
-                css_parts.append(f".{cls.strip()}")
-    
-    if css_parts:
-        selectors.append({
-            "type": "css_combined",
-            "selector": f"page.locator('{''.join(css_parts)}')",
-            "priority": "medium"
-        })
-    
-    return selectors
-
-def suggest_interactions(element: dict) -> list:
-    """Suggest appropriate Playwright interactions for an element"""
-    suggestions = []
-    
-    role = element.get('role', '').lower()
-    tag = element.get('tagName', '').lower()
-    element_type = element.get('type', '').lower()
-    
-    # Click interactions
-    if role in ['button', 'link', 'tab', 'menuitem'] or tag in ['button', 'a']:
-        suggestions.append("click()")
-        if role == 'link' or tag == 'a':
-            suggestions.append("click(button='middle')")  # Middle click for links
-    
-    # Form interactions
-    elif role == 'textbox' or tag == 'input':
-        if element_type in ['text', 'email', 'password', 'search']:
-            suggestions.append("fill('text')")
-            suggestions.append("type('text')")
-            suggestions.append("clear()")
-        elif element_type in ['checkbox', 'radio']:
-            suggestions.append("check()")
-            suggestions.append("uncheck()")
-            if element_type == 'radio':
-                suggestions.append("set_checked(true)")
-    
-    # Select interactions
-    elif role == 'combobox' or tag == 'select':
-        suggestions.append("select_option('value')")
-        suggestions.append("select_option(label='label')")
-    
-    # Hover interactions
-    if role in ['button', 'link', 'menuitem'] or tag in ['button', 'a']:
-        suggestions.append("hover()")
-    
-    # Focus interactions
-    if role in ['textbox', 'combobox', 'button'] or tag in ['input', 'select', 'button']:
-        suggestions.append("focus()")
-    
-    return suggestions
 
 # ========== CONFIGURABLE PARAMETERS ==========
 PHASE = 1
@@ -682,117 +54,6 @@ KNOWLEDGE_BASE_TYPE = os.getenv("KNOWLEDGE_BASE_TYPE", "graphrag")  # Type of kn
 os.makedirs(BROWSER_SESSIONS_DIR, exist_ok=True)
 
 
-
-
-
-
-
-def get_all_open_tabs(browser) -> list:
-    """Get information about all currently open tabs, excluding about:blank"""
-    tabs = []
-    try:
-        for page in browser.pages:
-            try:
-                # Skip about:blank tabs
-                if page.url == "about:blank":
-                    continue
-                    
-                try:
-                    tab_info = {
-                        'url': page.url,
-                        'title': page.title(),
-                        'domain': 'google.com' if 'google.com' in page.url.lower() else 'external',
-                        'page': page
-                    }
-                    tabs.append(tab_info)
-                except Exception as e:
-                    print(f"⚠️  Error getting tab info: {e}")
-                    # Skip this tab if we can't get its info
-                    continue
-            except Exception as e:
-                print(f"⚠️  Error accessing tab: {e}")
-                continue
-    except Exception as e:
-        print(f"⚠️  Error accessing browser pages: {e}")
-        return []
-    
-    return tabs
-
-def check_for_new_tabs(browser, previous_tab_count: int, previous_tab_urls: set) -> tuple[bool, list, int]:
-    """
-    Check if new tabs were opened and return info about them.
-    
-    Args:
-        browser: Browser context
-        previous_tab_count: Number of tabs from previous step
-        previous_tab_urls: Set of tab URLs from previous step
-        
-    Returns:
-        tuple: (has_new_tabs, new_tabs, current_tab_count)
-    """
-    current_tabs = get_all_open_tabs(browser)
-    current_tab_count = len(current_tabs)
-    current_tab_urls = {tab['url'] for tab in current_tabs}
-    
-    # Check if we have new tabs
-    if current_tab_count > previous_tab_count:
-        # Find new tabs (URLs that weren't in previous step)
-        new_tab_urls = current_tab_urls - previous_tab_urls
-        new_tabs = [tab for tab in current_tabs if tab['url'] in new_tab_urls]
-        
-        print(f"🆕 New tabs detected! Previous: {previous_tab_count}, Current: {current_tab_count}")
-        print(f"   New tabs: {[tab['domain'] for tab in new_tabs]}")
-        
-        return True, new_tabs, current_tab_count
-    else:
-        return False, [], current_tab_count
-
-def switch_to_new_tab(new_tabs: list, current_page) -> tuple[bool, object]:
-    """
-    Switch to the first new tab and return success status and new page object.
-    
-    Args:
-        new_tabs: List of new tab info dictionaries
-        current_page: Current page object
-        
-    Returns:
-        tuple: (success, new_page_object)
-    """
-    if not new_tabs:
-        return False, current_page
-    
-    try:
-        # Get the first new tab
-        new_tab = new_tabs[0]
-        new_page = new_tab['page']
-        
-        print(f"🔄 Switching to new tab: {new_tab['title']} ({new_tab['domain']})")
-        
-        # Wait for the new tab to be ready
-        print("⏳ Waiting for new tab to stabilize...")
-        new_page.wait_for_timeout(3000)  # 3 second delay
-        
-        # Bring the new tab to front
-        new_page.bring_to_front()
-        
-        # Wait a bit more after bringing to front
-        new_page.wait_for_timeout(2000)  # 2 second additional delay
-        
-        # Verify the tab is accessible
-        try:
-            new_page.wait_for_selector('body', timeout=5000)
-            print(f"✅ Successfully switched to: {new_tab['domain']}")
-            return True, new_page
-        except Exception as e:
-            print(f"⚠️  New tab not ready: {e}")
-            return False, current_page
-            
-    except Exception as e:
-        print(f"❌ Error switching to new tab: {e}")
-        return False, current_page
-
-
-
 def fetch_trajectory_nodes(
     instruction: str,
     max_results: int = 3,
@@ -808,791 +69,7 @@ def fetch_trajectory_nodes(
         max_context_length=max_context_length,
         kb_type=KNOWLEDGE_BASE_TYPE
     )
-
-def create_episode_directory(base_dir: str, eps_name: str) -> Dict[str, str]:
-    """Create directory structure for an episode."""
-    eps_dir = os.path.join(base_dir, eps_name)
-    dirs = {
-        'root': eps_dir,
-        'axtree': os.path.join(eps_dir, 'axtree'),
-        'images': os.path.join(eps_dir, 'images'),
-        'annotated_images': os.path.join(eps_dir, 'annotated_images'),
-        'user_message': os.path.join(eps_dir, 'user_message'),
-        'targeting_data': os.path.join(eps_dir, 'targeting_data'),
-        'gpt_summaries': os.path.join(eps_dir, 'gpt_summaries')
-    }
-    for dir_path in dirs.values():
-        os.makedirs(dir_path, exist_ok=True)
-    return dirs
-
-def create_trajectory_file(dirs: Dict[str, str]) -> None:
-    """Create an empty trajectory.json file with initial structure."""
-    trajectory_path = os.path.join(dirs['root'], 'trajectory.json')
-    with open(trajectory_path, 'w', encoding='utf-8') as f:
-        json.dump({}, f, indent=2, ensure_ascii=False)
-
-def create_error_log_file(dirs: Dict[str, str]) -> None:
-    """Create an empty error_log.json file with initial structure."""
-    error_log_path = os.path.join(dirs['root'], 'error_log.json')
-    with open(error_log_path, 'w', encoding='utf-8') as f:
-        json.dump({"playwright_errors": []}, f, indent=2, ensure_ascii=False)
-
-def update_playwright_error_log(dirs: Dict[str, str], step_idx: int, description: str, attempted_code: str, 
-                               error_message: str, successful_code: str = None, thought: str = None, 
-                               current_goal: str = None, all_failed_attempts: list = None) -> None:
-    """Update error_log.json with Playwright execution error information and solution."""
-    error_log_path = os.path.join(dirs['root'], 'error_log.json')
     
-    try:
-        with open(error_log_path, 'r', encoding='utf-8') as f:
-            error_log = json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        error_log = {"playwright_errors": []}
-    
-    # Check if we already have an error entry for this step
-    existing_error = None
-    for error in error_log["playwright_errors"]:
-        if error.get("step_index") == step_idx:
-            existing_error = error
-            break
-    
-    if existing_error:
-        # Update existing error entry
-        if all_failed_attempts:
-            # This is a successful solution - update with solution and all attempts
-            existing_error["successful_playwright_code"] = successful_code
-            existing_error["attempted_codes"] = all_failed_attempts
-            existing_error["final_error_message"] = error_message
-        else:
-            # This is another failed attempt - add to attempted_codes
-            if "attempted_codes" not in existing_error:
-                existing_error["attempted_codes"] = []
-            
-            attempt_entry = {
-                "attempt_number": len(existing_error["attempted_codes"]) + 1,
-                "code": attempted_code,
-                "error_message": error_message,
-                "thought": thought,
-                "description": description
-            }
-            existing_error["attempted_codes"].append(attempt_entry)
-    else:
-        # Create new error entry
-        error_entry = {
-            "step_index": step_idx,
-            "timestamp": datetime.now().isoformat(),
-            "current_goal": current_goal,
-            "attempted_codes": []
-        }
-        
-        # Add constant fields to main body if they don't change
-        if description:
-            error_entry["description"] = description
-        if thought:
-            error_entry["thought"] = thought
-        
-        # Add first attempt
-        attempt_entry = {
-            "attempt_number": 1,
-            "code": attempted_code,
-            "error_message": error_message
-        }
-        
-        # Add varying fields to attempt entry
-        if description and "description" not in error_entry:
-            attempt_entry["description"] = description
-        if thought and "thought" not in error_entry:
-            attempt_entry["thought"] = thought
-            
-        error_entry["attempted_codes"].append(attempt_entry)
-        
-        # If this is immediately successful, add the solution
-        if successful_code:
-            error_entry["successful_playwright_code"] = successful_code
-        
-        error_log["playwright_errors"].append(error_entry)
-    
-    with open(error_log_path, 'w', encoding='utf-8') as f:
-        json.dump(error_log, f, indent=2, ensure_ascii=False)
-
-
-
-def update_trajectory(dirs: Dict[str, str], step_idx: int, screenshot: str, axtree: str, action_code: str, action_description: str, page, user_message_file: str = None, llm_output=None, targeting_data_file: str = None, element_summary: str = None, annotation_id: str = None) -> None:
-    """Update trajectory.json with a new step."""
-    trajectory_path = os.path.join(dirs['root'], 'trajectory.json')
-    try:
-        with open(trajectory_path, 'r', encoding='utf-8') as f:
-            trajectory = json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        trajectory = {}
-    
-    # Get current page information with safety checks
-    try:
-        current_url = page.url if hasattr(page, 'url') else "Unknown"
-        page_title = page.title() if hasattr(page, 'title') else "Unknown"
-        open_pages = page.context.pages if hasattr(page, 'context') else []
-        open_pages_titles = [p.title() for p in open_pages] if open_pages else []
-        open_pages_urls = [p.url for p in open_pages] if open_pages else []
-    except Exception as e:
-        print(f"⚠️  Error getting page info in update_trajectory: {e}")
-        current_url = "Error getting URL"
-        page_title = "Error getting title"
-        open_pages_titles = []
-        open_pages_urls = []
-    
-    # Load targeting data and find element by annotation ID if available
-    element_data = None
-    if targeting_data_file and annotation_id and os.path.exists(targeting_data_file):
-        try:
-            with open(targeting_data_file, 'r', encoding='utf-8') as f:
-                targeting_data = json.load(f)
-            
-            # Find element by annotation_id
-            element_data = next((elem for elem in targeting_data if str(elem.get('annotation_id', '')) == str(annotation_id)), None)
-        except Exception as e:
-            print(f"⚠️ Error loading targeting data: {e}")
-    
-    # Extract action type and locator from the code
-    action_type = None
-    locator_code = None
-    action_output = None
-    
-    # Get thought from LLM output, fallback to derived thought if not available
-    thought = llm_output.get('thought', '') if llm_output else ''
-    
-    # Parse the action code to determine type and get element properties
-    if "page.goto" in action_code:
-        action_type = "goto"
-        url = action_code.split("page.goto(")[1].split(")")[0].strip('"\'')
-        action_output = {
-            "thought": thought,
-            "action": {
-                "url": url
-            },
-            "action_name": "goto"
-        }
-    elif ".click()" in action_code:
-        action_type = "click"
-        locator_code = action_code.split(".click()")[0]
-        
-        # Use targeting data if available, otherwise fall back to page evaluation
-        if element_data:
-            # Get role and name directly from targeting data
-            role = element_data.get('element_info', {}).get('role', '')
-            name = element_data.get('element_info', {}).get('name', '')
-            
-            # Get button name from targeting data
-            button_name = name or element_data.get('element_info', {}).get('id', '')
-            
-            if button_name:
-                thought = f'I need to click the "{button_name}" button.'
-            else:
-                thought = 'I need to click a button.'
-            
-            bbox = element_data.get('bounding_box', {})
-            action_output = {
-                "thought": thought,
-                "action": {
-                    "bid": "",
-                    "button": "left",
-                    "click_type": "single",
-                    "bbox": [
-                        bbox.get('x', 0),
-                        bbox.get('y', 0),
-                        bbox.get('width', 0),
-                        bbox.get('height', 0)
-                    ],
-                    "class": element_data.get('element_info', {}).get('class_name', ''),
-                    "id": element_data.get('element_info', {}).get('id', ''),
-                    "type": element_data.get('element_info', {}).get('tag_name', ''),
-                    "ariaLabel": element_data.get('element_info', {}).get('name', ''),
-                    "role": element_data.get('element_info', {}).get('role', ''),
-                    "value": element_data.get('element_info', {}).get('value', ''),
-                    "node_properties": {
-                        "role": role,
-                        "value": name
-                    }
-                },
-                "action_name": "click",
-                "annotation_id": annotation_id
-            }
-        else:
-            # Fallback to page evaluation if no targeting data available
-            try:
-                element_info = page.evaluate("""() => {
-                    const lastClicked = document.activeElement;
-                    if (!lastClicked) return null;
-                    const rect = lastClicked.getBoundingClientRect();
-                    return {
-                        bbox: {
-                            x: rect.x,
-                            y: rect.y,
-                            width: rect.width,
-                            height: rect.height
-                        },
-                        class: lastClicked.className,
-                        id: lastClicked.id,
-                        type: lastClicked.tagName.toLowerCase(),
-                        ariaLabel: lastClicked.getAttribute('aria-label'),
-                        role: lastClicked.getAttribute('role'),
-                        value: lastClicked.value
-                    };
-                }""")
-            except Exception as e:
-                print(f"⚠️  Error evaluating page in update_trajectory (click): {e}")
-                element_info = None
-            if element_info:
-                # Get role and name directly from page evaluation
-                role = element_info.get('role', '')
-                name = element_info.get('value', '')
-                
-                # Get button name from page evaluation
-                button_name = name or element_info.get('ariaLabel') or element_info.get('id') or ''
-                if button_name:
-                    thought = f'I need to click the "{button_name}" button.'
-                else:
-                    thought = 'I need to click a button.'
-                action_output = {
-                    "thought": thought,
-                    "action": {
-                        "bid": "",
-                        "button": "left",
-                        "click_type": "single",
-                        "bbox": [
-                            element_info['bbox']['x'],
-                            element_info['bbox']['y'],
-                            element_info['bbox']['width'],
-                            element_info['bbox']['height']
-                        ],
-                        "class": element_info.get('class', ''),
-                        "id": element_info.get('id', ''),
-                        "type": element_info.get('type', ''),
-                        "ariaLabel": element_info.get('ariaLabel', ''),
-                        "role": element_info.get('role', ''),
-                        "value": element_info.get('value', ''),
-                        "node_properties": {
-                            "role": role,
-                            "value": name
-                        }
-                    },
-                    "action_name": "click",
-                    "annotation_id": annotation_id
-                }
-    elif ".fill(" in action_code:
-        action_type = "type"
-        parts = action_code.split(".fill(")
-        locator_code = parts[0]
-        text = parts[1].split(")")[0].strip('"\'')
-        # Get the last focused input element
-        try:
-            element_info = page.evaluate("""() => {
-                const lastFocused = document.activeElement;
-                if (!lastFocused) return null;
-                const rect = lastFocused.getBoundingClientRect();
-                return {
-                    bbox: {
-                        x: rect.x,
-                        y: rect.y,
-                        width: rect.width,
-                        height: rect.height
-                    },
-                    class: lastFocused.className,
-                    id: lastFocused.id,
-                    type: lastFocused.tagName.toLowerCase(),
-                    ariaLabel: lastFocused.getAttribute('aria-label'),
-                    role: lastFocused.getAttribute('role'),
-                    value: lastFocused.value
-                };
-            }""")
-        except Exception as e:
-            print(f"⚠️  Error evaluating page in update_trajectory (fill): {e}")
-            element_info = None
-        if element_info:
-            action_output = {
-                "thought": thought,
-                "action": {
-                    "text": text
-                },
-                "action_name": "keyboard_type"
-            }
-    elif ".click()" in action_code and ".type(" in action_code:
-        # Handle combined click + type (common in fallback scenarios)
-        action_type = "type"
-        parts = action_code.split(".click()")
-        locator_code = parts[0]
-        
-        # Extract text to type
-        text = action_code.split(".type(")[1].split(")")[0].strip('"\'')
-        
-        # Use targeting data if available, otherwise fall back to page evaluation
-        if element_data:
-            bbox = element_data.get('bounding_box', {})
-            action_output = {
-                "thought": f'I need to click and type "{text}" into the input field.',
-                "action": {
-                    "bid": "",
-                    "button": "left",
-                    "click_type": "single",
-                    "bbox": [
-                        bbox.get('x', 0),
-                        bbox.get('y', 0),
-                        bbox.get('width', 0),
-                        bbox.get('height', 0)
-                    ],
-                    "class": element_data.get('element_info', {}).get('class_name', ''),
-                    "id": element_data.get('element_info', {}).get('id', ''),
-                    "type": element_data.get('element_info', {}).get('tag_name', ''),
-                    "ariaLabel": element_data.get('element_info', {}).get('name', ''),
-                    "role": element_data.get('element_info', {}).get('role', ''),
-                    "value": element_data.get('element_info', {}).get('value', ''),
-                    "text": text,
-                    "node_properties": {
-                        "role": element_data.get('element_info', {}).get('role', ''),
-                        "value": text
-                    }
-                },
-                "action_name": "type",
-                "annotation_id": annotation_id
-            }
-        else:
-            # Fallback to page evaluation if no targeting data available
-            try:
-                element_info = page.evaluate("""() => {
-                    const lastClicked = document.activeElement;
-                    if (!lastClicked) return null;
-                    const rect = lastClicked.getBoundingClientRect();
-                    return {
-                        bbox: {
-                            x: rect.x,
-                            y: rect.y,
-                            width: rect.width,
-                            height: rect.height
-                        },
-                        class: lastClicked.className,
-                        id: lastClicked.id,
-                        type: lastClicked.tagName.toLowerCase(),
-                        ariaLabel: lastClicked.getAttribute('aria-label'),
-                        role: lastClicked.getAttribute('role'),
-                        value: lastClicked.value
-                    };
-                }""")
-            except Exception as e:
-                print(f"⚠️  Error evaluating page in update_trajectory (fill): {e}")
-                element_info = None
-            if element_info:
-                action_output = {
-                    "thought": f'I need to click and type "{text}" into the input field.',
-                    "action": {
-                        "bid": "",
-                        "button": "left",
-                        "click_type": "single",
-                        "bbox": [
-                            element_info['bbox']['x'],
-                            element_info['bbox']['y'],
-                            element_info['bbox']['width'],
-                            element_info['bbox']['height']
-                        ],
-                        "class": element_info.get('class', ''),
-                        "id": element_info.get('id', ''),
-                        "type": element_info.get('type', ''),
-                        "ariaLabel": element_info.get('ariaLabel', ''),
-                        "role": element_info.get('role', ''),
-                        "value": element_info.get('value', ''),
-                        "text": text,
-                        "node_properties": {
-                            "role": element_info.get('role', ''),
-                            "value": text
-                        }
-                    },
-                    "action_name": "type",
-                    "annotation_id": annotation_id
-                }
-    elif "page.mouse.click(" in action_code:
-        action_type = "click"
-        
-        # Extract coordinates from the mouse click
-        coords_match = re.search(r'page\.mouse\.click\((\d+),\s*(\d+)\)', action_code)
-        if coords_match:
-            x, y = int(coords_match.group(1)), int(coords_match.group(2))
-            
-            # Use targeting data if available, otherwise fall back to page evaluation
-            if element_data:
-                # Get role and name directly from targeting data
-                role = element_data.get('element_info', {}).get('role', '')
-                name = element_data.get('element_info', {}).get('name', '')
-                
-                # Get button name from targeting data
-                button_name = name or element_data.get('element_info', {}).get('id', '')
-                
-                if button_name:
-                    thought = f'I need to click the "{button_name}" button.'
-                else:
-                    thought = 'I need to click a button.'
-                
-                bbox = element_data.get('bounding_box', {})
-                action_output = {
-                    "thought": thought,
-                    "action": {
-                        "bid": "",
-                        "button": "left",
-                        "click_type": "single",
-                        "bbox": [
-                            bbox.get('x', 0),
-                            bbox.get('y', 0),
-                            bbox.get('width', 0),
-                            bbox.get('height', 0)
-                        ],
-                        "class": element_data.get('element_info', {}).get('class_name', ''),
-                        "id": element_data.get('element_info', {}).get('id', ''),
-                        "type": element_data.get('element_info', {}).get('tag_name', ''),
-                        "ariaLabel": element_data.get('element_info', {}).get('name', ''),
-                        "role": element_data.get('element_info', {}).get('role', ''),
-                        "value": element_data.get('element_info', {}).get('value', ''),
-                        "coordinates": [x, y],
-                        "node_properties": {
-                            "role": role,
-                            "value": name
-                        }
-                    },
-                    "action_name": "click",
-                    "annotation_id": annotation_id
-                }
-            else:
-                # Fallback to page evaluation if no targeting data available
-                try:
-                    element_info = page.evaluate("""() => {
-                        const elementAtPoint = document.elementFromPoint(arguments[0], arguments[1]);
-                        if (!elementAtPoint) return null;
-                        const rect = elementAtPoint.getBoundingClientRect();
-                        return {
-                            bbox: {
-                                x: rect.x,
-                                y: rect.y,
-                                width: rect.width,
-                                height: rect.height
-                            },
-                            class: elementAtPoint.className,
-                            id: elementAtPoint.id,
-                            type: elementAtPoint.tagName.toLowerCase(),
-                            ariaLabel: elementAtPoint.getAttribute('aria-label'),
-                            role: elementAtPoint.getAttribute('role'),
-                            value: elementAtPoint.value
-                        };
-                    }""", x, y)
-                except Exception as e:
-                    print(f"⚠️  Error evaluating page in update_trajectory (mouse click): {e}")
-                    element_info = None
-                
-                if element_info:
-                    # Get role and name directly from page evaluation
-                    role = element_info.get('role', '')
-                    name = element_info.get('value', '')
-                    
-                    # Get button name from page evaluation
-                    button_name = name or element_info.get('ariaLabel') or element_info.get('id', '')
-                    
-                    if button_name:
-                        thought = f'I need to click the "{button_name}" button.'
-                    else:
-                        thought = 'I need to click a button.'
-                    
-                    action_output = {
-                        "thought": thought,
-                        "action": {
-                            "bid": "",
-                            "button": "left",
-                            "click_type": "single",
-                            "bbox": [
-                                element_info['bbox']['x'],
-                                element_info['bbox']['y'],
-                                element_info['bbox']['width'],
-                                element_info['bbox']['height']
-                            ],
-                            "class": element_info.get('class', ''),
-                            "id": element_info.get('id', ''),
-                            "type": element_info.get('type', ''),
-                            "ariaLabel": element_info.get('ariaLabel', ''),
-                            "role": element_info.get('role', ''),
-                            "value": element_info.get('value', ''),
-                            "coordinates": [x, y],
-                            "node_properties": {
-                                "role": role,
-                                "value": name
-                            }
-                        },
-                        "action_name": "click",
-                        "annotation_id": annotation_id
-                    }
-    elif ".dblclick()" in action_code:
-        action_type = "dblclick"
-        locator_code = action_code.split(".dblclick()")[0]
-        try:
-            element_info = page.evaluate("""() => {
-                const lastClicked = document.activeElement;
-                if (!lastClicked) return null;
-                const rect = lastClicked.getBoundingClientRect();
-                return {
-                    bbox: {
-                        x: rect.x,
-                        y: rect.y,
-                        width: rect.width,
-                        height: rect.height
-                    },
-                    class: lastClicked.className,
-                    id: lastClicked.id,
-                    type: lastClicked.tagName.toLowerCase(),
-                    ariaLabel: lastClicked.getAttribute('aria-label'),
-                    role: lastClicked.getAttribute('role'),
-                    value: lastClicked.value
-                };
-            }""")
-        except Exception as e:
-            print(f"⚠️  Error evaluating page in update_trajectory (dblclick): {e}")
-            element_info = None
-        if element_info:
-            action_output = {
-                "thought": thought,
-                "action": {
-                    "bid": "",
-                    "button": "left",
-                    "click_type": "double",
-                    "bbox": [
-                        element_info['bbox']['x'],
-                        element_info['bbox']['y'],
-                        element_info['bbox']['width'],
-                        element_info['bbox']['height']
-                    ],
-                    "node_properties": {
-                        "role": element_info.get('role', ''),
-                        "value": element_info.get('value', '')
-                    }
-                },
-                "action_name": "dblclick"
-            }
-    elif "page.scroll" in action_code:
-        action_type = "scroll"
-        action_output = {
-            "thought": thought,
-            "action": {},
-            "action_name": "scroll"
-        }
-    elif ".paste(" in action_code:
-        action_type = "paste"
-        locator_code = action_code.split(".paste(")[0]
-        try:
-            element_info = page.evaluate("""() => {
-                const lastFocused = document.activeElement;
-                if (!lastFocused) return null;
-                const rect = lastFocused.getBoundingClientRect();
-                return {
-                    bbox: {
-                        x: rect.x,
-                        y: rect.y,
-                        width: rect.width,
-                        height: rect.height
-                    },
-                    class: lastFocused.className,
-                    id: lastFocused.id,
-                    type: lastFocused.tagName.toLowerCase(),
-                    ariaLabel: lastFocused.getAttribute('aria-label'),
-                    role: lastFocused.getAttribute('role'),
-                    value: lastFocused.value
-                };
-            }""")
-        except Exception as e:
-            print(f"⚠️  Error evaluating page in update_trajectory (paste): {e}")
-            element_info = None
-        if element_info:
-            action_output = {
-                "thought": thought,
-                "action": {
-                    "bid": "",
-                    "bbox": [
-                        element_info['bbox']['x'],
-                        element_info['bbox']['y'],
-                        element_info['bbox']['width'],
-                        element_info['bbox']['height']
-                    ],
-                    "node_properties": {
-                        "role": element_info.get('role', ''),
-                        "value": element_info.get('value', '')
-                    }
-                },
-                "action_name": "paste"
-            }
-    elif "page.keyboard.press" in action_code:
-        action_type = "keypress"
-        key = action_code.split("page.keyboard.press(")[1].split(")")[0].strip('"\'')
-        try:
-            element_info = page.evaluate("""() => {
-                const lastFocused = document.activeElement;
-                if (!lastFocused) return null;
-                const rect = lastFocused.getBoundingClientRect();
-                return {
-                    bbox: {
-                        x: rect.x,
-                        y: rect.y,
-                        width: rect.width,
-                        height: rect.height
-                    },
-                    class: lastFocused.className,
-                    id: lastFocused.id,
-                    type: lastFocused.tagName.toLowerCase(),
-                    ariaLabel: lastFocused.getAttribute('aria-label'),
-                    role: lastFocused.getAttribute('role'),
-                    value: lastFocused.value
-                };
-            }""")
-        except Exception as e:
-            print(f"⚠️  Error evaluating page in update_trajectory (keyboard.press): {e}")
-            element_info = None
-        if element_info:
-            action_output = {
-                "thought": thought,
-                "action": {
-                    "key": key,
-                    "bid": "",
-                    "bbox": [
-                        element_info['bbox']['x'],
-                        element_info['bbox']['y'],
-                        element_info['bbox']['width'],
-                        element_info['bbox']['height']
-                    ],
-                    "node_properties": {
-                        "role": element_info.get('role', ''),
-                        "value": element_info.get('value', '')
-                    }
-                },
-                "action_name": "keypress"
-            }
-    
-    # Generate high-level action_str for the step
-    action_str = None
-    if "page.goto" in action_code:
-        url = action_code.split("page.goto(")[1].split(")")[0].strip('"\'')
-        action_str = f"goto(url='{url}')"
-    elif ".click()" in action_code:
-        bid = ""
-        button = "left"
-        if action_output and "action" in action_output:
-            bid = action_output["action"].get("bid", "")
-            button = action_output["action"].get("button", "left")
-        if bid or button:
-            action_str = f"click(bid='{bid}', button='{button}')"
-        else:
-            action_str = "click(...)"
-    elif ".fill(" in action_code:
-        text = action_code.split(".fill(")[1].split(")")[0].strip('"\'')
-        action_str = f"keyboard_type(text='{text}')"
-    elif ".dblclick()" in action_code:
-        action_str = "dblclick(...)"
-    elif "page.scroll" in action_code:
-        action_str = "scroll(...)"
-    elif ".paste(" in action_code:
-        action_str = "paste(...)"
-    elif "page.keyboard.press" in action_code:
-        key = action_code.split("page.keyboard.press(")[1].split(")")[0].strip('"\'')
-        action_str = f"keyboard_press(key='{key}')"
-    else:
-        action_str = action_code
-    
-    # Add new step
-    trajectory[str(step_idx + 1)] = {
-        "screenshot": os.path.basename(screenshot),
-        "axtree": os.path.basename(axtree),
-        "targeting_data": os.path.join('targeting_data', os.path.basename(targeting_data_file)) if targeting_data_file else None,
-        "user_message": os.path.join('user_message', os.path.basename(user_message_file)) if user_message_file else None,
-        "other_obs": {
-            "page_index": 0,
-            "url": current_url,
-            "open_pages_titles": open_pages_titles,
-            "open_pages_urls": open_pages_urls
-        },
-        "action": {
-            "action_str": action_str,
-            "playwright_code": action_code,
-            "action_description": action_description,
-            "action_output": action_output
-        },
-        "error": None,
-        "action_timestamp": time.time()
-    }
-    
-    with open(trajectory_path, 'w', encoding='utf-8') as f:
-        json.dump(trajectory, f, indent=2, ensure_ascii=False)
-
-def create_metadata(persona: str, url: str, orig_instruction: str, aug_instruction: str, 
-                   final_instruction: Optional[str], steps: list, success: bool, total_steps: int,
-                   runtime: float, total_tokens: int, page, eps_name: str) -> Dict[str, Any]:
-    """Create metadata dictionary."""
-    # Get viewport size
-    viewport = page.viewport_size
-    viewport_str = f"{viewport['width']}x{viewport['height']}" if viewport else "unknown"
-    
-    # Get browser context info
-    context = page.context
-    cookies_enabled = context.cookies() is not None
-    
-    return {
-        "goal": orig_instruction,
-        "eps_name": eps_name,
-        "task": {
-            "task_type": "calendar",
-            "steps": steps,
-            "instruction": {
-                "high_level": orig_instruction,
-                "mid_level": aug_instruction,
-                "low_level": final_instruction if final_instruction else aug_instruction
-            }
-        },
-        "start_url": url,
-        "browser_context": {
-            "os": os.uname().sysname.lower(),  # Get OS name
-            "viewport": viewport_str,
-            "cookies_enabled": cookies_enabled
-        },
-        "success": success,
-        "total_steps": total_steps,
-        "runtime_sec": runtime,
-        "total_tokens": total_tokens,
-        "phase": PHASE
-    }
-
-
-
-
-
-def write_user_message(user_message_file: str, goal: str, execution_history: list, page, tree, failed_codes: list = None):
-    """Write a user message file with goal, previous actions, current page, ax tree, and error codes."""
-    user_message_content = []
-    user_message_content.append(f"Goal: {goal}\n")
-    user_message_content.append("Previous Actions:")
-    if execution_history:
-        for i, act in enumerate(execution_history, 1):
-            user_message_content.append(f"  {i}. {act['step']} | Code: {act['code']}")
-    else:
-        user_message_content.append("  None")
-    user_message_content.append("")
-    # Safety check for page object
-    try:
-        page_title = page.title() if hasattr(page, 'title') else "Unknown"
-        page_url = page.url if hasattr(page, 'url') else "Unknown"
-        user_message_content.append(f"Current Page: {page_title} ({page_url})\n")
-    except Exception as e:
-        print(f"⚠️  Error getting page info: {e}")
-        user_message_content.append("Current Page: Unknown (Error getting page info)\n")
-    user_message_content.append("AX Tree:")
-    user_message_content.append(json.dumps(tree, indent=2, ensure_ascii=False))
-    user_message_content.append("")
-    user_message_content.append("Error Codes:")
-    if failed_codes:
-        for err in failed_codes:
-            user_message_content.append(f"  {err}")
-    else:
-        user_message_content.append("  None")
-    with open(user_message_file, 'w', encoding='utf-8') as f:
-        f.write('\n'.join(user_message_content))
-
 def generate_trajectory_loop(user_data_dir, chrome_path, phase, start_idx, end_idx, email: Optional[str] = None, password: Optional[str] = None, search_context: bool = False):
     phase_file = os.path.join(RESULTS_DIR, f"instructions_phase{phase}.json")
     try:
@@ -2278,158 +755,7 @@ def main():
         for future in futures:
             future.result()  # Wait for all to finish
 
-def generate_trajectory_html(dirs: Dict[str, str], metadata: Dict[str, Any]) -> None:
-    """Generate an HTML visualization of the trajectory."""
-    trajectory_path = os.path.join(dirs['root'], 'trajectory.json')
-    html_path = os.path.join(dirs['root'], 'trajectory.html')
-    
-    try:
-        with open(trajectory_path, 'r', encoding='utf-8') as f:
-            trajectory = json.load(f)
-    except (FileNotFoundError, json.JSONDecodeError):
-        print("❌ Error loading trajectory.json")
-        return
 
-    # Instruction Table
-    instructions = metadata['task']['instruction']
-    steps = metadata['task']['steps']
-    html_content = f"""<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Visualization of Trajectory</title>
-    <style>
-        body {{ font-family: Arial, sans-serif; line-height: 1.6; margin: 0; padding: 20px; background-color: #f5f5f5; }}
-        .container {{ max-width: 1200px; margin: 0 auto; background-color: white; padding: 20px; border-radius: 8px; box-shadow: 0 2px 4px rgba(0,0,0,0.1); }}
-        h1, h2 {{ color: #333; border-bottom: 2px solid #eee; padding-bottom: 10px; }}
-        .step {{ border: 1px solid #ddd; margin: 20px 0; padding: 15px; border-radius: 5px; background-color: white; }}
-        .step-header {{ display: flex; justify-content: space-between; align-items: center; margin-bottom: 10px; }}
-        .step-number {{ font-size: 1.2em; font-weight: bold; color: #2196F3; }}
-        .step-content {{ display: grid; grid-template-columns: 1fr 1fr; gap: 20px; }}
-        .screenshot {{ max-width: 100%; border: 1px solid #ddd; border-radius: 4px; }}
-        .collapsible {{ background-color: #eee; color: #444; cursor: pointer; padding: 10px; width: 100%; border: none; text-align: left; outline: none; font-size: 1em; margin-top: 5px; }}
-        .active, .collapsible:hover {{ background-color: #ccc; }}
-        .content {{ padding: 0 18px; display: none; overflow: auto; background-color: #f9f9f9; border-radius: 0 0 5px 5px; }}
-        pre {{ white-space: pre-wrap; word-break: break-word; }}
-        table.instruction-table {{ width: 100%; border-collapse: collapse; margin-bottom: 20px; }}
-        table.instruction-table th, table.instruction-table td {{ border: 1px solid #ddd; padding: 8px; }}
-        table.instruction-table th {{ background: #f0f0f0; text-align: left; }}
-        .steps-list {{ margin: 0; padding-left: 20px; }}
-        .steps-list li {{ margin-bottom: 4px; }}
-        .step-details-label {{ font-weight: bold; margin-top: 10px; }}
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>{metadata['eps_name']} ({metadata['task'].get('task_type','')})</h1>
-        <p><strong>GPT Summaries:</strong> <a href="gpt_summaries/">View all GPT summary files</a></p>
-        <h2>Instructions</h2>
-        <table class="instruction-table">
-            <tr><th>level</th><th>instruction</th></tr>
-            <tr><td><em>high_level</em></td><td>{html.escape(instructions.get('high_level',''))}</td></tr>
-            <tr><td><em>mid_level</em></td><td>{html.escape(instructions.get('mid_level',''))}</td></tr>
-            <tr><td><em>low_level</em></td><td>{html.escape(instructions.get('low_level',''))}</td></tr>
-            <tr><td><em>steps</em></td><td><ul class="steps-list">{''.join(f'<li>{html.escape(str(s))}</li>' for s in steps)}</ul></td></tr>
-        </table>
-        <h2>Trajectory Steps</h2>
-"""
-
-    for step_num, step_data in trajectory.items():
-        screenshot_path = os.path.join('images', step_data['screenshot'])
-        user_message_path = step_data.get('user_message')
-        targeting_data_path = step_data.get('targeting_data')
-        
-        user_message_content = ""
-        if user_message_path:
-            user_message_full_path = os.path.join(dirs['root'], user_message_path)
-            try:
-                with open(user_message_full_path, 'r', encoding='utf-8') as umf:
-                    user_message_content = html.escape(umf.read())
-            except Exception:
-                user_message_content = "[Could not load user message]"
-        else:
-            user_message_content = "[No user message]"
-        action = step_data['action']
-        action_output = action.get('action_output', {})
-        # Fix: check if action_output is a dict before calling .get()
-        thought = html.escape(action_output.get('thought', '')) if isinstance(action_output, dict) else ''
-        action_str = html.escape(action.get('action_str', ''))
-        action_description = html.escape(action.get('action_description', ''))
-        # System message: use a field if available, else placeholder
-        system_message = step_data.get('system_message', 'System message for this step (placeholder)')
-        # Element output: pretty print action_output['action'] if available
-        element_output = ''
-        if isinstance(action_output, dict) and 'action' in action_output:
-            element_output = json.dumps(action_output['action'], indent=2, ensure_ascii=False)
-            element_output = html.escape(element_output)
-        else:
-            element_output = '[No element output]'
-        # LLM output: show Playwright code for this step
-        playwright_code = action.get('playwright_code', '')
-        llm_output_str = html.escape(playwright_code) if playwright_code else 'No Playwright code for this step.'
-
-        # Get annotated screenshot path if available
-        annotated_screenshot_path = None
-        if step_data.get('targeting_data'):
-            # Try to find corresponding annotated screenshot
-            step_num_int = int(step_num)
-            annotated_filename = f"annotated_screenshot_{step_num_int:03d}.png"
-            annotated_screenshot_path = os.path.join('annotated_images', annotated_filename)
-        
-        html_content += f"""
-        <div class="step">
-            <div class="step-header">
-                <span class="step-number">Step {step_num}</span>
-            </div>
-            <div class="step-content">
-                <div>
-                    <img src="{screenshot_path}" alt="Step {step_num} Screenshot" class="screenshot">
-                    {f'<br><br><strong>Annotated Version:</strong><br><img src="{annotated_screenshot_path}" alt="Step {step_num} Annotated Screenshot" class="screenshot">' if annotated_screenshot_path else ''}
-                </div>
-                <div>
-                    <div class="step-details-label">Thought</div>
-                    <div>{thought}</div>
-                    <div class="step-details-label">Action Code</div>
-                    <div><pre>{llm_output_str}</pre></div>
-                    <div class="step-details-label">Action Description</div>
-                    <div>{action_description}</div>
-                    <button class="collapsible">System Message</button>
-                    <div class="content"><pre>{system_message}</pre></div>
-                    <button class="collapsible">User Message</button>
-                    <div class="content"><pre>{user_message_content}</pre></div>
-                    <button class="collapsible">Element Output</button>
-                    <div class="content"><pre>{element_output}</pre></div>
-                    <button class="collapsible">LLM Output</button>
-                    <div class="content"><pre>{llm_output_str}</pre></div>
-                </div>
-            </div>
-        </div>
-        """
-
-    html_content += """
-    </div>
-    <script>
-    document.addEventListener('DOMContentLoaded', function() {
-      var coll = document.getElementsByClassName('collapsible');
-      for (var i = 0; i < coll.length; i++) {
-        coll[i].addEventListener('click', function() {
-          this.classList.toggle('active');
-          var content = this.nextElementSibling;
-          if (content.style.display === 'block') {
-            content.style.display = 'none';
-          } else {
-            content.style.display = 'block';
-          }
-        });
-      }
-    });
-    </script>
-</body>
-</html>"""
-
-    with open(html_path, 'w', encoding='utf-8') as f:
-        f.write(html_content)
 
 if __name__ == "__main__":
     main()
